physic: tweak constants to force godoc
This follows up on c9566d5acafe9f5217. I observed that if three quarter of the
constants have the type, it's enough to regroup the constants under the type in
godoc. The problem is that in practice this is quite confusing to read, so at
that point it's better to just add the type everywhere even if it's noisy.