Add GitHub actions
Take inspiration from the workflows used in gohci and bootstrap but
disable many checks because it's going to take a while to fix them all.

Update go generate to make it work without prerequisites preinstalled.

Remove the corresponding checks from .travis.yml. There are still checks
that haven't been transitioned, this will be done in a follow up.