Start over the import path again
I hadn't realized that v0 is not a valid suffix and must not be used.
That's really unfortunate because I'd prefer to start right away with
versioned imports, it'd make the follow up sed much easier.

It's especially annoying that go test ./... did not notify of this
problem and I only realized after I pushed v0.1.0. When I forcibly
requested it I got an error. I deleted the tag for now.

Ran:
  git ls-files -- . | xargs -L 1 sed -i 's#conn/v0#conn#g'